Enh: Allow loss percentages less than 1 and also a global command line option to specify that packets should be lost at a given percentage - provided by Charles P. Wright from IBM Research
git-svn-id: svn://svn.code.sf.net/p/sipp/code/sipp/trunk@185 60913f6b-29f6-41b8-8391-1ad2d2d6c344

+++ scenario.hpp
@@ -123,7 +123,7 @@ public:
   int            stop_rtd;
   bool           repeat_rtd;
   int		 counter;
-  int            lost;
+  double         lost;
   int            crlf;
   unsigned int   next;
   int            test;
